refact: Merge recurring contributions queries
