+++ components/recurring-contributions/graphql/queries.js
@@ -1,0 +1,47 @@
+import { gqlV2 } from '../../../lib/graphql/helpers';
+
+export const recurringContributionsQuery = gqlV2/* GraphQL */ `
+  query RecurringContributions($slug: String) {
+    account(slug: $slug) {
+      id
+      slug
+      name
+      type
+      settings
+      imageUrl
+      orders(filter: OUTGOING, status: [ACTIVE, CANCELLED]) {
+        totalCount
+        nodes {
+          id
+          paymentMethod {
+            id
+          }
+          amount {
+            value
+            currency
+          }
+          status
+          frequency
+          tier {
+            id
+            name
+          }
+          totalDonations {
+            value
+            currency
+          }
+          toAccount {
+            id
+            slug
+            name
+            type
+            description
+            tags
+            imageUrl
+            settings
+          }
+        }
+      }
+    }
+  }
+`;
